changed namespace so we can have online javadocs
