--- src/client/java/com/fofr/FracturedModClient.java
@@ -1,10 +1,0 @@
-package com.fofr;
-
-import net.fabricmc.api.ClientModInitializer;
-
-public class FracturedModClient implements ClientModInitializer {
-	@Override
-	public void onInitializeClient() {
-		// This entrypoint is suitable for setting up client-specific logic, such as rendering.
-	}
-}
