--- src/client/java/com/fofr/FracturedModDataGenerator.java
@@ -1,11 +1,0 @@
-package com.fofr;
-
-import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
-import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
-
-public class FracturedModDataGenerator implements DataGeneratorEntrypoint {
-	@Override
-	public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
-
-	}
-}
