+++ src/client/java/io/github/bobisawesome07/FracturedModClient.java
@@ -1,0 +1,10 @@
+package io.github.bobisawesome07;
+
+import net.fabricmc.api.ClientModInitializer;
+
+public class FracturedModClient implements ClientModInitializer {
+	@Override
+	public void onInitializeClient() {
+		// This entrypoint is suitable for setting up client-specific logic, such as rendering.
+	}
+}
