+++ src/client/java/io/github/bobisawesome07/FracturedModDataGenerator.java
@@ -1,0 +1,11 @@
+package io.github.bobisawesome07;
+
+import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
+import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
+
+public class FracturedModDataGenerator implements DataGeneratorEntrypoint {
+	@Override
+	public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
+
+	}
+}
