--- src/main/java/com/fofr/FractureModDataGenerator.java
@@ -1,21 +1,0 @@
-package com.fofr;
-
-import com.fofr.world.dimension.ModDimensions;
-import com.fofr.datagen.ModWorldGenerator;
-import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
-import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
-import net.minecraft.registry.RegistryBuilder;
-import net.minecraft.registry.RegistryKeys;
-
-public class FractureModDataGenerator implements DataGeneratorEntrypoint{
-
-    @Override
-    public void onInitializeDataGenerator(FabricDataGenerator generator) {
-        FabricDataGenerator.Pack pack = generator.createPack();
-        pack.addProvider(ModWorldGenerator::new);
-    }
-    @Override
-    public void buildRegistry(RegistryBuilder registryBuilder){
-		registryBuilder.addRegistry(RegistryKeys.DIMENSION_TYPE, ModDimensions::bootstrapType);
-    }
-}
