--- src/main/java/com/fofr/datagen/ModWorldGenerator.java
@@ -1,24 +1,0 @@
-package com.fofr.datagen;
-
-import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
-import net.fabricmc.fabric.api.datagen.v1.provider.FabricDynamicRegistryProvider;
-import net.minecraft.registry.RegistryKeys;
-import net.minecraft.registry.RegistryWrapper;
-
-import java.util.concurrent.CompletableFuture;
-
-public class ModWorldGenerator extends FabricDynamicRegistryProvider {
-    public ModWorldGenerator(FabricDataOutput output, CompletableFuture<RegistryWrapper.WrapperLookup> registriesFuture) {
-        super(output, registriesFuture);
-    }
-
-    @Override
-    protected void configure(RegistryWrapper.WrapperLookup registries, Entries entries) {
-        entries.addAll(registries.getWrapperOrThrow(RegistryKeys.DIMENSION_TYPE));
-    }
-
-    @Override
-    public String getName() {
-        return "World Gen";
-    }
-}
