--- src/main/java/com/fofr/entity/ModEntities.java
@@ -1,4 +1,0 @@
-package com.fofr.entity;
-
-public class ModEntities {
-}
