+++ src/main/java/io/github/bobisawesome07/FractureModDataGenerator.java
@@ -1,0 +1,21 @@
+package io.github.bobisawesome07;
+
+import io.github.bobisawesome07.world.dimension.ModDimensions;
+import io.github.bobisawesome07.datagen.ModWorldGenerator;
+import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
+import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
+import net.minecraft.registry.RegistryBuilder;
+import net.minecraft.registry.RegistryKeys;
+
+public class FractureModDataGenerator implements DataGeneratorEntrypoint{
+
+    @Override
+    public void onInitializeDataGenerator(FabricDataGenerator generator) {
+        FabricDataGenerator.Pack pack = generator.createPack();
+        pack.addProvider(ModWorldGenerator::new);
+    }
+    @Override
+    public void buildRegistry(RegistryBuilder registryBuilder){
+		registryBuilder.addRegistry(RegistryKeys.DIMENSION_TYPE, ModDimensions::bootstrapType);
+    }
+}
