+++ src/main/java/io/github/bobisawesome07/entity/ModEntities.java
@@ -1,0 +1,4 @@
+package io.github.bobisawesome07.entity;
+
+public class ModEntities {
+}
